rename data compressor
