+++ config/grid_search_override.yaml
@@ -3,7 +3,7 @@ run_name_content: [[msg_size, batch_size, pred_hidden_sizes, lr]]
 
 batch_size: [50]
 lr: [1e-2, 1e-3]
-data_compressor_name: [KME]
+data_encoder_name: [KME]
 
 mlp_1_dim: [[50]]  # [[50], [100], [500], [1000]]
 #msg_size: ${mlp_1_dim[0]}
