+++ src/model/data_encoder/conv_one_by_one_block.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
 
-from src.model.data_compressor.DataEncoder import DataEncoder
+from src.model.data_encoder.data_encoder import DataEncoder
 
 
 class ConvOneByOneBlock(nn.Module, DataEncoder):
