+++ src/model/data_encoder/fspool.py
@@ -1,7 +1,7 @@
 import torch
 from torch import nn as nn
 
-from src.model.data_compressor.DataEncoder import DataEncoder
+from src.model.data_encoder.data_encoder import DataEncoder
 from src.model.mlp import MLP
